src/main/java/com/azs/beefygainz/exercise/controller/ExerciseController.java
package com.azs.beefygainz.exercise.controller;

import com.azs.beefygainz.exercise.model.Exercise;
import com.azs.beefygainz.exercise.service.ExerciseService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/exercises")
public class ExerciseController {

    private final ExerciseService exerciseService;

    @Autowired
    public ExerciseController(ExerciseService exerciseService) {
        this.exerciseService = exerciseService;
    }

    @GetMapping("")
    public List<Exercise> getExercises(@RequestHeader("Auth-Token") String userId) {
        return exerciseService.findAllByUserId(userId);
    }

    @PostMapping("")
    @ResponseStatus(HttpStatus.CREATED)
    public Exercise saveExercise(@RequestHeader("Auth-Token") String userId, @RequestBody Exercise exercise) {
        return exerciseService.save(exercise, userId);
    }
}
